refactor: reorganize SDK exports and update import paths
- Move modules/ to chain/ for better organization
- Move metadata files to dedicated metadata/ directory
- Move address.ts to types/ directory
- Remove re-exports from index.browser.ts
- Add modular exports to package.json (alphabetically ordered):
  - ./cached-queries
  - ./chain
  - ./constants
  - ./metadata
- Update all consuming packages to use direct module imports
- Fix unused variable warning by prefixing with underscore
- Add agent-metadata exports to metadata index

This completes the SDK export refactoring to use a more modular structure
where consumers import directly from specific modules rather than the main
package export.
