+++ .vscode/settings.json
@@ -103,6 +103,7 @@
     "prettiercache",
     "pubk",
     "rainbowkit",
+    "redelegated",
     "rootnet",
     "rustie",
     "shadcn",
