+++ apps/torus-allocator/src/app/_components/agent-card/agent-card-content.tsx
@@ -3,7 +3,7 @@
 import { CardContent } from "@torus-ts/ui/components/card";
 import { Separator } from "@torus-ts/ui/components/separator";
 import { useQueryAgentMetadata } from "~/hooks/use-agent-metadata";
-import type { AgentMetadataResult } from "@torus-network/sdk";
+import type { AgentMetadataResult } from "@torus-network/sdk/metadata";
 import { SkeletonAgentCardContent } from "./agent-card-skeleton-loader";
 
 interface AgentCardContentProps {
