+++ apps/torus-allocator/src/hooks/use-weekly-usd.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 
-import { CONSTANTS } from "@torus-network/sdk";
+import { CONSTANTS } from "@torus-network/sdk/constants";
 
 import {
   useGetTorusPrice,
