+++ apps/torus-cache/src/data.ts
@@ -1,13 +1,16 @@
 import type { ApiPromise } from "@polkadot/api";
-import type { LastBlock, StakeData } from "@torus-network/sdk";
+import SuperJSON from "superjson";
+
+import type { LastBlock } from "@torus-network/sdk";
 import {
   queryLastBlock,
   queryStakeIn,
   queryStakeOut,
 } from "@torus-network/sdk";
+import type { StakeData } from "@torus-network/sdk/cached-queries";
 import { BasicLogger } from "@torus-network/torus-utils/logger";
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
-import SuperJSON from "superjson";
+
 import { setup } from "./server";
 import { sleep } from "./utils";
 
@@ -206,4 +209,4 @@ export async function updateStakeDataLoop() {
     await sleep(5000);
   }
 }
-export { stakeOutData, stakeFromData };
+export { stakeFromData, stakeOutData };
