+++ apps/torus-portal/src/app/(pages)/(permission-graph)/_components/graph-sheet/graph-sheet-agent-card/agent-card-container.tsx
@@ -2,7 +2,7 @@
 
 import { memo, useCallback, useEffect, useMemo, useState } from "react";
 
-import { fetchAgentMetadata } from "@torus-network/sdk";
+import { fetchAgentMetadata } from "@torus-network/sdk/metadata";
 import { smallAddress } from "@torus-network/torus-utils/subspace";
 import { tryAsync } from "@torus-network/torus-utils/try-catch";
