+++ apps/torus-portal/src/app/(pages)/(permission-graph)/_components/permission-graph-utils.tsx
@@ -1,4 +1,4 @@
-import { CONSTANTS } from "@torus-network/sdk";
+import { CONSTANTS } from "@torus-network/sdk/constants";
 
 import type {
   CachedAgentData,
