+++ apps/torus-wallet/src/hooks/useRewardInterval.ts
@@ -1,4 +1,4 @@
-import { CONSTANTS } from "@torus-network/sdk";
+import { CONSTANTS } from "@torus-network/sdk/constants";
 import { useEffect, useState } from "react";
 import { useWallet } from "~/context/wallet-provider";
