+++ apps/torus-worker/src/common/index.ts
@@ -8,11 +8,11 @@ import type {
   Proposal,
 } from "@torus-network/sdk";
 import {
-  CONSTANTS,
   queryAgentApplications,
   queryLastBlock,
   queryProposals,
 } from "@torus-network/sdk";
+import { CONSTANTS } from "@torus-network/sdk/constants";
 import { BasicLogger } from "@torus-network/torus-utils/logger";
 import { tryAsync } from "@torus-network/torus-utils/try-catch";
