+++ examples/example-client/src/index.ts
@@ -1,24 +1,50 @@
-import * as torus from "@torus-network/sdk";
+import { z } from "zod";
 
-import { ApiPromise, WsProvider } from "@polkadot/api";
+import { Agent } from "@torus-network/sdk";
 
-async function connectToChainRpc(wsEndpoint: string) {
-  const wsProvider = new WsProvider(wsEndpoint);
-  const api = await ApiPromise.create({ provider: wsProvider });
-  if (!api.isConnected) {
-    throw new Error("API not connected");
-  }
-  console.log("API connected");
-  return api;
-}
+const agent = new Agent({
+  agentKey: "5FgfC2DY4yreEWEughz46RZYQ8oBhHVqD9fVq6gV89E6z4Ea", // Your agent's SS58 address
+  port: 3000,
+  docs: {
+    info: {
+      title: "Alice Memory Agent",
+      version: "1.0.0",
+    },
+  },
+});
 
-async function main() {
-  const api = await connectToChainRpc("wss://api.testnet.torus.network");
+// Define a simple endpoint
+agent.method(
+  "hello",
+  {
+    input: z.object({
+      name: z.string().min(1).max(50),
+    }),
+    output: {
+      ok: {
+        description: "Greeting response",
+        schema: z.object({
+          message: z.string(),
+          timestamp: z.number(),
+        }),
+      },
+      err: {
+        description: "Error response",
+        schema: z.object({
+          error: z.string(),
+        }),
+      },
+    },
+  },
+  async (input, context) => {
+    return {
+      ok: {
+        message: `Hello ${input.name}!`,
+        timestamp: Date.now(),
+      },
+    };
+  },
+);
 
-  const agents = await torus.queryAgents(api);
-  console.log(agents);
-
-  await api.disconnect();
-}
-
-await main();
+// Start the server
+agent.run();
