+++ packages/query-provider/src/hooks.ts
@@ -16,19 +16,15 @@ import type {
   PermissionId,
   Proposal,
   SS58Address,
-  StakeData,
   VoteWithStake,
 } from "@torus-network/sdk";
 import {
-  CONSTANTS,
-  fetchCustomMetadata,
   processVotesAndStakes,
   queryAccountsNotDelegatingVotingPower,
   queryAgentApplications,
   queryAgents,
   queryBlockEmission,
   queryBurnValue,
-  queryCachedStakeOut,
   queryDaoTreasuryAddress,
   queryExtFee,
   queryFreeBalance,
@@ -53,6 +49,10 @@ import {
   queryUnrewardedProposals,
   queryWhitelist,
 } from "@torus-network/sdk";
+import { queryCachedStakeOut } from "@torus-network/sdk/cached-queries";
+import type { StakeData } from "@torus-network/sdk/cached-queries";
+import { CONSTANTS } from "@torus-network/sdk/constants";
+import { fetchCustomMetadata } from "@torus-network/sdk/metadata";
 import { BasicLogger } from "@torus-network/torus-utils/logger";
 import { tryAsync } from "@torus-network/torus-utils/try-catch";
 import type { ListItem, Nullish } from "@torus-network/torus-utils/typing";
