+++ packages/torus-provider/src/_components/send-transaction.tsx
@@ -5,7 +5,7 @@ import type { InjectedAccountWithMeta } from "@polkadot/extension-inject/types";
 import type { DispatchError } from "@polkadot/types/interfaces";
 import { u8aToHex } from "@polkadot/util";
 
-import { CONSTANTS } from "@torus-network/sdk";
+import { CONSTANTS } from "@torus-network/sdk/constants";
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
 
 import { toast } from "@torus-ts/ui/hooks/use-toast";
