+++ packages/torus-sdk-ts/src/chain/_common.ts
@@ -5,8 +5,8 @@ import type {
   SubmittableExtrinsic,
 } from "@polkadot/api/types";
 import type { StorageKey } from "@polkadot/types";
-import type { AccountId, Extrinsic } from "@polkadot/types/interfaces";
-import type { Codec } from "@polkadot/types/types";
+import type { AccountId, Extrinsic, Header } from "@polkadot/types/interfaces";
+import type { Codec, IU8a } from "@polkadot/types/types";
 import type { z, ZodTypeAny } from "zod";
 
 import { assert_error } from "@torus-network/torus-utils";
@@ -14,7 +14,8 @@ import type { Result } from "@torus-network/torus-utils/result";
 import { makeErr, makeOk } from "@torus-network/torus-utils/result";
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
 
-import { sb_balance } from "../index.browser.js";
+import type { Blocks } from "../index.browser.js";
+import { sb_balance, sb_blocks } from "../index.browser.js";
 
 export type Api = ApiDecoration<"promise"> | ApiPromise;
 
@@ -178,7 +179,7 @@ interface FeeInfo {
 /**
  * @deprecated Not working
  */
-export async function queryExtFeeInfo(
+export async function brokenQueryExtFeeInfo(
   api: Api,
   ext: Extrinsic,
 ): Promise<Result<FeeInfo, SbQueryError>> {
@@ -226,4 +227,48 @@ export async function queryExtFee(
   }
   const fee = payInfo.partialFee.toBigInt();
   return makeOk({ fee });
+}
+export async function queryLastBlock(api: ApiPromise): Promise<LastBlock> {
+  const [headerError, blockHeader] = await tryAsync(api.rpc.chain.getHeader());
+  if (headerError !== undefined) {
+    console.error("Error getting block header:", headerError);
+    throw headerError;
+  }
+
+  const [apiError, apiAtBlock] = await tryAsync(api.at(blockHeader.hash));
+  if (apiError !== undefined) {
+    console.error("Error getting API at block:", apiError);
+    throw apiError;
+  }
+
+  const [parseError, blockNumber] = trySync(() =>
+    sb_blocks.parse(blockHeader.number),
+  );
+  if (parseError !== undefined) {
+    console.error("Error parsing block number:", parseError);
+    throw parseError;
+  }
+
+  const [hashError, blockHashHex] = trySync(() => blockHeader.hash.toHex());
+  if (hashError !== undefined) {
+    console.error("Error converting block hash to hex:", hashError);
+    throw hashError;
+  }
+
+  const lastBlock = {
+    blockHeader,
+    blockNumber,
+    blockHash: blockHeader.hash,
+    blockHashHex,
+    apiAtBlock,
+  };
+
+  return lastBlock;
+}
+export interface LastBlock {
+  blockHeader: Header;
+  blockNumber: Blocks;
+  blockHash: IU8a;
+  blockHashHex: `${string}`;
+  apiAtBlock: Api;
 }
