+++ packages/torus-sdk-ts/src/chain/subspace.ts
@@ -7,7 +7,7 @@ import type { z } from "zod";
 
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
 
-import type { SS58Address } from "../address.js";
+import type { SS58Address } from "../types/address.js";
 import type { Balance } from "../types/index.js";
 import {
   sb_address,
