+++ packages/torus-sdk-ts/src/evm.ts
@@ -7,8 +7,10 @@ import {
 import { assert } from "tsafe";
 import type { Chain, WalletClient } from "viem";
 import { encodeFunctionData } from "viem";
-import type { SS58Address } from "./address.js";
+
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
+
+import type { SS58Address } from "./types/address.js";
 
 export { waitForTransactionReceipt } from "@wagmi/core";
