+++ packages/torus-sdk-ts/src/metadata/index.ts
@@ -1,0 +1,2 @@
+export * from "./metadata.js";
+export * from "./agent-metadata.js";
