+++ packages/torus-sdk-ts/src/types/index.ts
@@ -1,4 +1,9 @@
 export * from "./zod.js";
 export * from "./base.js";
 
+export * from "./address.js";
 export * from "./namespace/index.js";
+
+export * from "./events.js";
+
+// TODO: organize directory
