+++ packages/torus-sdk-ts/src/types/zod.ts
@@ -18,7 +18,7 @@ import { z } from "zod";
 
 import type { Option } from "@torus-network/torus-utils";
 
-import { SS58_SCHEMA } from "../address.js";
+import { SS58_SCHEMA } from "./address.js";
 
 export { sb_enum } from "./sb_enum.js";
